requirements.txt
bm3d==4.0.1
bm4d==4.2.3
Brotli 
certifi==2023.7.22
cffi
charset-normalizer
contourpy==1.1.1
cvxpy==1.3.2
cycler==0.12.0
ecos==2.0.12
einops==0.7.0
et-xmlfile==1.1.0
filelock
fonttools==4.43.0
fsspec==2023.9.2
idna 
imageio==2.31.5
importlib-resources==6.1.0
Jinja2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe
matplotlib==3.7.3
monai==1.3.0
mpmath
networkx 
numpy 
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.4
osqp==0.6.3
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pycparser
pyparsing==3.1.1
PySocks 
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
qdldl==0.1.7.post0
requests 
scikit-image==0.21.0
scipy==1.10.1
scs==3.2.3
six==1.16.0
sympy
tifffile==2023.7.10
torch==2.2.0
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.1
triton==2.2.0
typing_extensions
tzdata==2024.1
urllib3
visdom==0.2.4
websocket-client==1.6.3
zipp==3.17.0
